Move reducers and actions into services
