--- client/src/middlewares/middlewares.tsx
@@ -1,11 +1,0 @@
-export default [
-  function thunkMiddleware({ dispatch, getState }: any) {
-    return function(next: any) {
-      return function(action: any) {
-        return typeof action === "function"
-          ? action(dispatch, getState)
-          : next(action);
-      };
-    };
-  }
-];
