+++ client/src/services/toaster/actions.tsx
@@ -1,6 +1,6 @@
 import { IToastProps } from "@blueprintjs/core";
 
-import { IAction } from "../types";
+import { IAction } from "../../types";
 
 export const addToast = (props: IToastProps): IAction => {
   return {
