+++ client/src/services/toaster/reducer.tsx
@@ -1,5 +1,5 @@
-import { IToastProps, IToaster, Position, Toaster } from "@blueprintjs/core";
-import { ISimpleAction } from "../types";
+import { Position, Toaster } from "@blueprintjs/core";
+import { ISimpleAction } from "../../types";
 
 const initialState: any = Toaster.create({
   className: "recipe-toaster",
